Move char joiner registry to browser layer
