+++ src/browser/renderer/CharacterJoinerRegistry.ts
@@ -4,7 +4,7 @@
  */
 
 import { IBufferLine, ICellData, CharData } from 'common/Types';
-import { ICharacterJoinerRegistry, ICharacterJoiner } from './Types';
+import { ICharacterJoinerRegistry, ICharacterJoiner } from 'browser/renderer/Types';
 import { AttributeData } from 'common/buffer/AttributeData';
 import { WHITESPACE_CELL_CHAR, Content } from 'common/buffer/Constants';
 import { CellData } from 'common/buffer/CellData';
@@ -90,7 +90,7 @@ export class CharacterJoinerRegistry implements ICharacterJoinerRegistry {
     }
 
     const line = this._bufferService.buffer.lines.get(row);
-    if (line.length === 0) {
+    if (!line || line.length === 0) {
       return [];
     }
